start the rewrite

+++ tests/support/mod.rs
@@ -1,13 +1,11 @@
-use vt100;
-
 pub trait TestHelpers {
-    fn assert_process(&mut self, s: &[u8]) -> u64;
+    fn assert_process(&mut self, s: &[u8]) -> usize;
 }
 
 impl TestHelpers for vt100::Screen {
-    fn assert_process(&mut self, s: &[u8]) -> u64 {
+    fn assert_process(&mut self, s: &[u8]) -> usize {
         let ret = self.process(s);
-        assert_eq!(ret, s.len() as u64);
+        assert_eq!(ret, s.len());
         ret
     }
 }
